Add files via upload
This dashboard contains data set of a Company that sells cars and the dataset contains the sales he company made on cars sold from various manufacturing company within a year. It contains information like Car Manufacturers, Horse Power , Engine , Fuel Capacity and lots more. The dashboard is insightful and talks about the performance of car sales in a year.